cloud_run/product_scraper/requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
lxml==4.9.3
webdriver-manager==4.0.1
python-dotenv==1.0.0
fake-useragent==1.4.0
google-cloud-storage==2.10.0
flask==2.3.3
gunicorn==21.2.0
undetected-chromedriver==3.5.4
